LayoutTests/imported/w3c/web-platform-tests/mediacapture-fromelement/ended-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.
CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: The operation is not supported.

Harness Error (FAIL), message = Unhandled rejection: The operation is not supported.

PASS video-only
FAIL ended assert_true: expected true got false
PASS audio-only
FAIL ended 1 assert_true: expected true got false
PASS video+audio
FAIL ended 2 assert_true: expected true got false
